♻️ Refactor: Add tooltip to sign up for and change folder structure
