+++ client/src/ui/Tooltip/Tooltip.module.scss
@@ -3,7 +3,7 @@
 }
 
 .content {
-  @apply p-2 inline-block absolute text-sm text-neutral-100 whitespace-nowrap bg-neutral-800 rounded shadow-lg shadow-black/[0.5] z-50;
+  @apply px-2 py-1 inline-block absolute text-sm text-white/[0.8]  whitespace-nowrap bg-neutral-900 rounded-lg shadow-lg shadow-black/[0.5] z-50;
 }
 
 // Top
